langchain_langgraph/pyproject.toml
[project]
name = "langchain-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "grandalf>=0.8",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.2",
    "langchain-groq>=1.1.2",
    "langchain-tavily>=0.2.18",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
]
